Reactivate check for already added hyperplanes

+++ src/DualSolver.h
@@ -28,7 +28,7 @@ public:
     void checkDualSolutionCandidates();
 
     void addGeneratedHyperplane(const Hyperplane& hyperplane);
-    bool hasHyperplaneBeenAdded(size_t hash, int constraintIndex);
+    bool hasHyperplaneBeenAdded(double hash, int constraintIndex);
 
     std::vector<GeneratedHyperplane> generatedHyperplanes;
     std::vector<std::pair<VectorInteger, VectorInteger>> generatedIntegerCuts;
